feat: wip NexusSDK class
