--- src/buyCover/index.ts
@@ -1,1 +1,0 @@
-export * from './calculatePremiumWithCommissionAndSlippage';
