+++ src/constants/cover.ts
