+++ src/constants/index.ts
@@ -1,5 +1,5 @@
 const { allPrivateProductsIds } = require('./privateProducts') as { allPrivateProductsIds: number[] };
 
 export { allPrivateProductsIds };
-export * from './buyCover';
+export * from './cover';
 export * from './products';
