+++ src/index.ts
@@ -1,6 +1,6 @@
 import * as deployments from '@nexusmutual/deployments';
 
-import * as buyCover from './buyCover';
+import * as cover from './cover';
 import * as constants from './constants';
 import * as ipfs from './ipfs';
 import * as quote from './quote';
@@ -9,6 +9,7 @@ import * as types from './types';
 import productTypesData from '../generated/product-types.json';
 import productsData from '../generated/products.json';
 import * as generatedTypes from '../generated/types';
+import { NexusSDK } from './nexus-sdk';
 
 const nexusSdk = {
   ...deployments,
@@ -16,11 +17,12 @@ const nexusSdk = {
   productTypes: productTypesData,
   ...generatedTypes,
   ...swap,
-  ...buyCover,
+  ...cover,
   ...types,
   ...quote,
   ...ipfs,
   ...constants,
+  NexusSDK,
 };
 
 // Re-export everything from the deployments package (e.g. `addresses` and `abis`)
@@ -35,7 +37,7 @@ export * from '../generated/types';
 
 export * from './swap';
 
-export * from './buyCover';
+export * from './cover';
 
 export * from './types';
 
@@ -44,5 +46,7 @@ export * from './ipfs';
 export * from './quote';
 
 export * from './constants';
+
+export { NexusSDK } from './nexus-sdk';
 
 export default nexusSdk;
