+++ src/swap/index.ts
@@ -1,8 +1,2 @@
-export * from './calculateExactNxmForEth';
-export * from './calculateExactEthForNxm';
-export * from './calculateNxmForExactEth';
-export * from './calculateEthForExactNxm';
 export * from './reserves.type';
-export * from './calculatePriceImpactA';
-export * from './calculatePriceImpactB';
-export * from './calculateSpotPrice';
+export * from './Swap';
