+++ src/types/cover.ts
@@ -1,6 +1,6 @@
 import { ApiResponse } from './api';
 import { Address, FloatString, IntString, Integer } from './data';
-import { CoverAsset } from '../constants/buyCover';
+import { CoverAsset } from '../constants/cover';
 
 export type BuyCoverParams = {
   coverId: Integer;
