+++ src/types/index.ts
@@ -2,3 +2,4 @@ export * from './api';
 export * from './cover';
 export * from './data';
 export * from './ipfs';
+export * from './sdk';
